make surveys load faster
