+++ met-web/src/components/survey/create/CloneOptions.tsx
@@ -2,7 +2,7 @@ import React, { useContext, useEffect, useState } from 'react';
 import { Grid, TextField, Stack, Autocomplete, Typography } from '@mui/material';
 import { CreateSurveyContext } from './CreateSurveyContext';
 import { useNavigate } from 'react-router-dom';
-import { cloneSurvey, getSurveysPage } from 'services/surveyService';
+import { cloneSurvey, fetchSurveys, getSurveysPage } from 'services/surveyService';
 import { getEngagements } from 'services/engagementService';
 import { useAppDispatch } from 'hooks';
 import { openNotification } from 'services/notificationService/notificationSlice';
@@ -15,27 +15,29 @@ export type EngagementParams = {
     engagementId: string;
 };
 
+interface CloneOptions {
+    availableSurveys: Survey[] | null;
+    loadingSurveys: boolean;
+}
+
 const PAGE = 1;
 
 const PAGE_SIZE = 2000;
 
 const SORT_ORDER = 'asc';
 
-const CloneOptions = () => {
+const CloneOptions = ({ availableSurveys, loadingSurveys }: CloneOptions) => {
     const navigate = useNavigate();
     const dispatch = useAppDispatch();
     const searchParams = new URLSearchParams(location.search);
     const engagementId = searchParams.get('engagementId');
     const [selectedSurvey, setSelectedSurvey] = useState<Survey | null>(null);
     const [selectedEngagement, setSelectedEngagement] = useState<Engagement | null>(null);
-    const [loadingSurveys, setLoadingSurveys] = useState(true);
     const [loadingEngagements, setLoadingEngagements] = useState(true);
     const [isSaving, setIsSaving] = useState(false);
     const {
         surveyForm,
         handleSurveyFormChange,
-        availableSurveys,
-        setAvailableSurveys,
         availableEngagements,
         setAvailableEngagements,
         isDisclaimerChecked,
@@ -69,21 +71,6 @@ const CloneOptions = () => {
         return '';
     };
 
-    const handleFetchSurveys = async (page: number, size: number, sort_order: 'asc' | 'desc' | undefined) => {
-        try {
-            const fetchedSurveys = await getSurveysPage({
-                page: page,
-                size: size,
-                sort_order: sort_order,
-                exclude_hidden: true,
-            });
-            setAvailableSurveys(fetchedSurveys.items);
-            setLoadingSurveys(false);
-        } catch (error) {
-            dispatch(openNotification({ severity: 'error', text: 'Error occurred while fetching available surveys' }));
-        }
-    };
-
     const handleFetchEngagements = async (page: number, size: number, sort_order: 'asc' | 'desc' | undefined) => {
         try {
             const fetchedEngagements = await getEngagements({
@@ -99,10 +86,6 @@ const CloneOptions = () => {
             );
         }
     };
-
-    useEffect(() => {
-        handleFetchSurveys(PAGE, PAGE_SIZE, SORT_ORDER);
-    }, []);
 
     useEffect(() => {
         if (!availableEngagements) {
